Move `catalog` subcommands to `cli/catalog.py`
